Added a new observation service so that it can be used by different entities
Updated auth steps for different resources
Made a few modifications to the unit tests now that auth is more stringent
