+++ src/main/resources/application.properties
@@ -79,14 +79,27 @@ quarkus.oidc.roles.source=accesstoken
 #quarkus.oidc.groups-claim=groups
 quarkus.oidc.roles.role-claim-path=groups
 
-# ---- Basic Auth only for /admin ----
-quarkus.http.auth.permission.admin.paths=/admin/*
+# ---- Auth ordering ----
+# TAP endpoint (no auth)
+quarkus.http.auth.permission.tap.paths=/archive/tap/*
+quarkus.http.auth.permission.tap.policy=permit
+
+# Basic Auth for /admin
+quarkus.http.auth.permission.admin.paths=/archive/admin/*
 quarkus.http.auth.permission.admin.policy=authenticated
 quarkus.http.auth.permission.admin.auth-mechanism=basic
 
-# ---- OIDC default for all endpoints ----
+# ---- OIDC default for all other endpoints ----
 quarkus.http.auth.permission.oidc.paths=/*
 quarkus.http.auth.permission.oidc.policy=authenticated
+quarkus.http.auth.permission.oidc.auth-mechanism=oidc
+
+# Disable security in test mode where required
+# Due to certain unit tests using direct HTTP calls, tokens cannot be generated
+%test.quarkus.http.auth.permission.collections.paths=/archive/collections,/archive/collections/*
+%test.quarkus.http.auth.permission.collections.policy=permit
+%test.quarkus.http.auth.permission.search.paths=/archive/search,/archive/search/*
+%test.quarkus.http.auth.permission.search.policy=permit
 
 # For validating the token
 # value MUST match the iss property in the token
@@ -166,7 +179,7 @@ testdata.seed.enabled=true
 
 #Basic Auth for admin user
 %dev.basic.auth.admin.username=admin
-basic.auth.role.username=${AUTH_ROLE:dev}
+basic.auth.admin.username=${AUTH_ROLE:dev}
 %dev.basic.auth.admin.password=aReallyLongPasswordForTestingOnly
 basic.auth.admin.password=${ADMIN_PASSWORD:dev}
