+++ src/test/java/org/uksrc/archive/ObservationResourceTest.java
@@ -278,6 +278,7 @@ public class ObservationResourceTest {
     @Test
     @DisplayName("Add an observation with a single artifact and check the response is the same.")
     @TestSecurity(user = "testuser", roles = {TEST_READER_ROLE, TEST_WRITER_ROLE})
+    @Transactional
     public void testGettingArtifactObservation() {
         //One plane with one artifact
         Observation obs1 = createArtifactObservation(OBSERVATION1, COLLECTION1, nonResolvableArtifactUri);
